Split features into feature.c
