+++ src/main/config/config.c
@@ -72,9 +72,9 @@
 #include "fc/runtime_config.h"
 
 #include "config/config.h"
-
 #include "config/config_profile.h"
 #include "config/config_master.h"
+#include "config/feature.h"
 
 #ifndef DEFAULT_RX_FEATURE
 #define DEFAULT_RX_FEATURE FEATURE_RX_PARALLEL_PWM
@@ -167,7 +167,6 @@ size_t custom_flash_memory_address = 0;
 
 master_t masterConfig;                 // master config struct with data independent from profiles
 profile_t *currentProfile;
-static uint32_t activeFeaturesLatch = 0;
 
 static uint8_t currentControlRateProfileIndex = 0;
 controlRateConfig_t *currentControlRateProfile;
@@ -1172,41 +1171,6 @@ void handleOneshotFeatureChangeOnRestart(void)
     if (feature(FEATURE_ONESHOT125) && !featureConfigured(FEATURE_ONESHOT125)) {
         delay(ONESHOT_FEATURE_CHANGED_DELAY_ON_BOOT_MS);
     }
-}
-
-void latchActiveFeatures()
-{
-    activeFeaturesLatch = masterConfig.enabledFeatures;
-}
-
-bool featureConfigured(uint32_t mask)
-{
-    return masterConfig.enabledFeatures & mask;
-}
-
-bool feature(uint32_t mask)
-{
-    return activeFeaturesLatch & mask;
-}
-
-void featureSet(uint32_t mask)
-{
-    masterConfig.enabledFeatures |= mask;
-}
-
-void featureClear(uint32_t mask)
-{
-    masterConfig.enabledFeatures &= ~(mask);
-}
-
-void featureClearAll()
-{
-    masterConfig.enabledFeatures = 0;
-}
-
-uint32_t featureMask(void)
-{
-    return masterConfig.enabledFeatures;
 }
 
 void persistentFlagClearAll()
