+++ src/main/fc/mw.c
@@ -86,6 +86,7 @@
 #include "config/config.h"
 #include "config/config_profile.h"
 #include "config/config_master.h"
+#include "config/feature.h"
 
 // June 2013     V2.2-dev
