+++ src/main/flight/mixer.c
@@ -59,6 +59,7 @@
 
 #include "config/config.h"
 #include "config/config_profile.h"
+#include "config/feature.h"
 
 //#define MIXER_DEBUG
