+++ src/main/io/beeper.c
@@ -41,6 +41,7 @@
 #include "fc/runtime_config.h"
 
 #include "config/config.h"
+#include "config/feature.h"
 
 #include "io/beeper.h"
