+++ src/main/io/serial_cli.c
@@ -91,6 +91,7 @@
 #include "config/config.h"
 #include "config/config_profile.h"
 #include "config/config_master.h"
+#include "config/feature.h"
 
 #include "common/printf.h"
