+++ src/main/io/serial_msp.c
@@ -88,6 +88,7 @@
 #include "config/config.h"
 #include "config/config_profile.h"
 #include "config/config_master.h"
+#include "config/feature.h"
 
 #include "blackbox/blackbox.h"
