+++ src/main/rx/pwm.c
@@ -33,6 +33,7 @@
 #include "drivers/pwm_rx.h"
 
 #include "config/config.h"
+#include "config/feature.h"
 
 #include "rx/rx.h"
 #include "rx/pwm.h"
