+++ src/main/sensors/battery.c
@@ -29,6 +29,7 @@
 #include "fc/runtime_config.h"
 
 #include "config/config.h"
+#include "config/feature.h"
 
 #include "sensors/battery.h"
