+++ src/main/telemetry/mavlink.c
@@ -74,6 +74,7 @@
 
 #include "config/config_profile.h"
 #include "config/config_master.h"
+#include "config/feature.h"
 
 #include "fc/mw.h"
